pyproject.toml
[tool.black]
include = '\.pyi?$'

[tool.poetry]
name = "ss-practice"
version = "0.1"
description = "Skyward Sword Practice"
authors = []

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
nlzss11 = "^1.5"
PyYAML = "^6.0"
PySide6 = "^6.4.3"
lark = "^1.1.5"
toml = "^0.10.2"
pyqtdarktheme = "^2.1.0"
pyclip = "^0.7.0"
opencv-python = "^4.7.0.68"
numpy = "^1.24.2"
disc-riider-py = "==0.3.1"
pyelf2rel = ">=1.0.9"

[tool.poetry.group.build.dependencies]
pyinstaller = "^5.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
